// src/models/public/PlaceActionReportQuantities.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { AccountsAccountId } from './Accounts';
import type { PlaceActionReportsPlaceActionReportId } from './PlaceActionReports';
import type { UnitsUnitId } from './Units';

/** Identifier type for public.place_action_report_quantities */
export type PlaceActionReportQuantitiesPlaceActionReportQuantityId = string & { __brand: 'public.place_action_report_quantities' };

/**
 * Represents the table public.place_action_report_quantities
 * Quantities of place action reports
 */
export default interface PlaceActionReportQuantities {
  place_action_report_quantity_id: PlaceActionReportQuantitiesPlaceActionReportQuantityId;

  /** redundant account_id enhances data safety */
  account_id: AccountsAccountId | null;

  place_action_report_id: PlaceActionReportsPlaceActionReportId | null;

  unit_id: UnitsUnitId | null;

  /** Used for integer quantities */
  quantity_integer: number | null;

  /** Used for numeric quantities */
  quantity_numeric: number | null;

  /** Used for text quantities */
  quantity_text: string | null;

  label: string | null;

  created_at: Date;

  updated_at: Date;

  updated_by: string | null;
}

/**
 * Represents the initializer for the table public.place_action_report_quantities
 * Quantities of place action reports
 */
export interface PlaceActionReportQuantitiesInitializer {
  /** Default value: uuid_generate_v7() */
  place_action_report_quantity_id?: PlaceActionReportQuantitiesPlaceActionReportQuantityId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  place_action_report_id?: PlaceActionReportsPlaceActionReportId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer quantities */
  quantity_integer?: number | null;

  /** Used for numeric quantities */
  quantity_numeric?: number | null;

  /** Used for text quantities */
  quantity_text?: string | null;

  label?: string | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;

  updated_by?: string | null;
}

/**
 * Represents the mutator for the table public.place_action_report_quantities
 * Quantities of place action reports
 */
export interface PlaceActionReportQuantitiesMutator {
  place_action_report_quantity_id?: PlaceActionReportQuantitiesPlaceActionReportQuantityId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  place_action_report_id?: PlaceActionReportsPlaceActionReportId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer quantities */
  quantity_integer?: number | null;

  /** Used for numeric quantities */
  quantity_numeric?: number | null;

  /** Used for text quantities */
  quantity_text?: string | null;

  label?: string | null;

  created_at?: Date;

  updated_at?: Date;

  updated_by?: string | null;
}